# movie-ratings/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Movie Ratings API",
    "version": "1.0.0",
    "description": ""
  },
  "host": "movierating-wakw.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/movieRating/{id}": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "example": "any"
                },
                "movieType": {
                  "example": "any"
                },
                "movieTitle": {
                  "example": "any"
                },
                "watchDate": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                },
                "movieGenre": {
                  "example": "any"
                },
                "movieReleaseYear": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}